Clean up --version, don't require a DISPLAY and display the server version too.
version.c was removed, seemed a bit excessive for the 20 lines of code.
--version is integrated separate from the other commands now, checked before
opening the display. xinput now prints its own version in all cases, even if
the display is unavailable. If the display is available, it prints the
server version too. Example output:

$> xinput --version
xinput version 1.4.99.3
XI version on server: 2.0

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>