list: only check the last bit in the device mode.
This works around a bug in X servers 1.7.x, 1.8.0 and 1.8.1 where the device
mode could sometimes be binary OR'd with the OutOfProximity flag. The result
was a value of 0b10 for relative and 0b11 for Absolute, both of which were
interpreted as relative by this code.

Affected is only the XIQueryDevice call, not the XListInputDevices call.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>